Use np.load and np.save for stable data storage
According to the NumPy documentation:

Do not rely on the combination of tofile and fromfile for data storage, as the binary files generated are are not platform independent. In particular, no byte-order or data-type information is saved. Data can be stored in the platform independent .npy format using save and load instead.

I gave an example using copy-on-write memory mapping, which should save memory in the common case without surprising people by writing changes back to disk.